fix(components): add JS export for interior left nav
Fixes #9.

This change also moves the bootstrap code out of index.js,
like https://github.com/carbon-design-system/carbon-components/pull/489 does.
