+++ src/globals/js/components.js
@@ -1,0 +1,10 @@
+export {
+  default as DetailPageHeader,
+} from '../../components/detail-page-header/detail-page-header';
+export {
+  default as InteriorLeftNav,
+} from '../../components/interior-left-nav/interior-left-nav';
+export { default as LeftNav } from '../../components/unified-header/left-nav';
+export {
+  default as ProfileSwitcher,
+} from '../../components/unified-header/profile-switcher';
