+++ src/globals/js/settings.js
@@ -1,0 +1,14 @@
+/**
+ * Settings.
+ * @exports CarbonComponents.settings
+ * @type Object
+ * @property {boolean} [disableAutoInit]
+ *   Disables automatic instantiation of components.
+ *   By default (`CarbonComponents.disableAutoInit` is `false`),
+ *   carbon-components attempts to instantiate components automatically
+ *   by searching for elements with `data-component-name` (e.g. `data-loading`) attribute
+ *   or upon DOM events (e.g. clicking) on such elements.
+ *   See each components' static `.init()` methods for details.
+ */
+const settings = {};
+export default settings;
